เชื่อม backend

+++ backend/app/services/ml_service.py
@@ -51,11 +51,11 @@ class MLService:
         img = img.resize((224, 224))
         arr = np.array(img, dtype=np.float32) / 255.0
         # Normalize with ImageNet mean/std
-        mean = np.array([0.485, 0.456, 0.406])
-        std  = np.array([0.229, 0.224, 0.225])
+        mean = np.array([0.485, 0.456, 0.406], dtype=np.float32)
+        std = np.array([0.229, 0.224, 0.225], dtype=np.float32)
         arr = (arr - mean) / std
         # NCHW format
-        arr = arr.transpose(2, 0, 1)[np.newaxis, :]
+        arr = arr.transpose(2, 0, 1)[np.newaxis, :].astype(np.float32)
         return arr
 
     def predict(self, image_bytes: bytes) -> dict:
